divide each enitity into a single file
