+++ server/tag.go
@@ -2,11 +2,36 @@ package main
 
 import (
 	"log"
+	"time"
 
 	"github.com/gin-gonic/gin"
 	"github.com/google/uuid"
 	_ "github.com/lib/pq"
 )
+
+var tagTable = `
+
+CREATE EXTENSION IF NOT EXISTS "uuid-ossp";
+
+CREATE TABLE IF NOT EXISTS tag (
+	id uuid UNIQUE,
+	name text,
+	user_id uuid,
+	created_at timestamp with time zone,
+	updated_at timestamp with time zone
+)`
+
+type Tag struct {
+	ID        string    `db:"id" json:"id"`
+	Name      string    `db:"name" json:"name"`
+	UserID    string    `db:"user_id" json:"user_id"`
+	CreatedAt time.Time `db:"created_at" json:"created_at"`
+	UpdatedAt time.Time `db:"updated_at" json:"updated_at"`
+}
+
+type TagInput struct {
+	Name string `json:"name" binding:"required"`
+}
 
 func insertTag(c *gin.Context) {
 	var newTag TagInput
